split chess into classes
